+++ elmyra.ip.access.epo/CHANGES.rst
@@ -9,6 +9,7 @@ development
 - middleware: add alternative cql parser implementation based on pyparsing, with doctests
 - middleware: make pyparsing cql parser unicode aware, tweak and cleanup things
 - middleware: add more inline doctests to pyparsing cql parser
+- middleware: make cql parser understand neighbourhood term operators
 
 0.28.0
 ======